refactor: extract 10 command groups from cli.ts into register functions
Closes #160

Reduces cli.ts from 940 to 585 lines (38% reduction) by moving command
group registrations into their respective command files, following the
existing registerXCommand(program) pattern used by trigger, approval,
autonomous, deploy, eval, and orchestrate commands.

Extracted groups: memory, sessions, session, goal, kpi, progress,
feedback, learn, env, exec.


Agent: engineering/issue-solver
Squad: engineering
Trigger: manual
Model: claude-opus-4-6
